--- UI/css/style.css
@@ -1,179 +1,0 @@
-body{
-    background-color: black;
-   
-    background-image: url(../assets/nature.jpg);
-    background-repeat:no-repeat;
-    background-size: cover;
-    width:100vw;
-    height: 10vh;
-    background-attachment: fixed;
-    background-position: top;
-    }
-
-
-    
-    .logo{
-    height: 10px auto;
-    width:15%;
-    }
-    nav ul{
-    list-style-type: none;
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    }
-    nav li{
-    flex: 1 0 auto;
-    }
-    nav li a{
-    display:block;
-    padding:10px;
-    font-size: 25px;
-    text-decoration: none;
-    text-align: center;
-    color:black;
-    transition: background 2s;
-    text-transform: uppercase;
-    }
-   
-    nav{
-    background-color:rgb(246, 243, 250);
-    
-    }
-    
-    nav li a:hover{
-        color:black;
-    background-color: rgba(230,122,0,0,3);
-    }
-    .content{
-    
-    display:flex;
-    text-align: top;
-    text-align: center;
-    justify-content: center;
-    flex-direction: colomn;
-
-    }
-    .main-content{
-        
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        text-align: center;
-        height: 20vh;
-        width: 100vw;
-    }
-    .articles{
-      display:flex; 
-      flex-wrap: wrap; 
-      flex-direction: row;
-      justify-content: flex-end;
-     
-    }
-    .article{
-        flex: 1 1 10px;
-        justify-content: flex-end;
-    }
-    .content h1{
-    font-size: 25px;
-    margin-left:180px;
-    
-
-    }
-    .content-b{
-      height: 10vh;
-      width: 100vw;
-      text-align: center;
-      display: flex;
-      flex-direction: row;
-      
-    }
-    .content-b p{
-      font-size: 20px;
-      color: antiquewhite;
-      text-align: center;
-    }
-
-  .content-b a{
-    text-decoration: none;
-
-  }
-
-    
-    .footer{
-        display:flex;
-        background-color: white;
-        height:30vh;
-        width:100vw;
-        flex-direction:column;
-        justify-content: center;
-        text-align: center;
-    }
-   
-    .main-content2{
-        
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        text-align: center;
-    }
-    .articles{
-      display:flex; 
-      flex-wrap: wrap; 
-      flex-direction: row;
-      justify-content: flex-end;
-     
-    }
-    .article{
-        flex: 1 1 10px;
-        justify-content: flex-end;
-    }
-    .kiki{
-        width:100vw;
-        height:50vh;
-         background-color:rgb(233, 243, 250); 
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        text-align: center;
-     
-      }
-      * {box-sizing:border-box}
-
-/* Slideshow container */
-.slideshow-container {
-  max-width: 1000px;
-  position: relative;
-  margin: auto;
-}
-
-/* Hide the images by default */
-.mySlides {
-  display: none;
-
-}
-
-.footer1{
-  display:flex;
-  background-color: white;
-  height:80vh;
-  width:100vw;
-  flex-direction:column;
-  justify-content: center;
-  text-align: center;
-  
-    
-
-        
-    
-   
-}
-.footer2{
-  display:flex;
-  background-color: white;
-  height:30vh;
-  width:100vw;
-  flex-direction:column;
-  justify-content: center;
-  text-align: center;
-}
