--- UI/html/entries.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="X-UA-Compatible" content="ie=edge">
-<link rel="stylesheet" href="../CSS/style.css">
-<title>Document</title>
-<style>
-body {font-family: Arial, Helvetica, sans-serif;}
-form {border: 3px solid #f1f1f1;}
-
-    
-.logo{
-    height: 60%;
-    width:15%;
-    }
-    
-
-    nav li a{
-    display:block;
-    padding:10px;
-    font-size: 25px;
-    text-decoration: none;
-    text-align: center;
-    color:black;
-    transition: background 2s;
-    }
-   
-    nav{
-    background-color:rgb(246, 243, 250);
-    
-    }
-    .nav-menu{
-    flex-grow: 1;
-    }
-    nav li a:hover{
-        color:black;
-    background-color: rgba(230,122,0,0,3);
-    }
-    .cont{
-  display:flex;
-  background-color: white;
-  height:100vh;
-  width:100vw;
-  flex-direction:column;
-  justify-content: top;
-  text-align: center;
-  margin: bottom;
-    }
-
-   .menus{
-    display:flex;
-    height: 50vh;
-   }
-  menus li{
-    padding: 20px;
-  }
-
-
-</style>
-</head>
-<body>
-    <header class="header">
-
-                <nav>
-                <ul class="nav-menu">
-                <img class="logo"src="../assets/logo.png">
-                
-                <li> <a href="ui.html" > HOME</a> </li>
-                <li> <a href="log.html" > logout</a> </li>
-                
-                </ul>
-                </nav>
-                </header> 
-                <div class="entries">
-                  <marquee direction="right"><h5> WELCOME TO YOUR PAGE</h5></marquee> <br>
-                <ent>
-<ul class="menus">
-<li> <a href="signup.html" > your account</a> </li> 
-<li> <a href="newj.html" > start new journal</a> </li>
-<li> <a href="#" > view entries</a> </li>
-<li> <a href="#" > modify entries</a> </li>
-<li> <a href="#" > settings</a> </li>
-</ul>
-</ent>
-
-                </div>
-                
-                </body>
