--- UI/html/newj.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="X-UA-Compatible" content="ie=edge">
-<link rel="stylesheet" href="../CSS/style.css">
-<title>Document</title>
-</head>
-<body>
-
-    <h1> start your journal here</h1>
-    <form action="" method="">
-        <textarea rows="30"cols="50">
-
-
-        </textarea><br><br>
-        <button type="save">save<a href ="#"></a></button>
-        <button type="update">update<a href ="#"></a></button>
-        <button type="cancel">cancel<a href ="#"></a></button>
-    </form>
-
-    </body>
