--- UI/html/signup.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html>
-<style>
-body {font-family: Arial, Helvetica, sans-serif;}
-* {box-sizing: border-box}
-
-
-input[type=text], input[type=password] {
-  width: 100%;
-  padding: 15px;
-  margin: 5px 0 22px 0;
-  display: inline-block;
-  border: none;
-  background: #f1f1f1;
-}
-
-input[type=text]:focus, input[type=password]:focus {
-  background-color: #ddd;
-  outline: none;
-}
-
-hr {
-  border: 1px solid #f1f1f1;
-  margin-bottom: 25px;
-}
-
-
-button {
-  background-color: skyblue;
-  color: white;
-  padding: 14px 20px;
-  margin: 8px 0;
-  border: none;
-  cursor: pointer;
-  width: 100%;
-  opacity: 0.9;
-}
-
-button:hover {
-  opacity:1;
-}
-
-
-.cancelbtn {
-  padding: 14px 20px;
-  background-color: black;
-}
-
-
-.cancelbtn, .signupbtn {
-  float: left;
-  width: 100%;
-}
-
-
-.container {
-  padding: 16px;
-}
-
-
-.clearfix::after {
-  content: "";
-  clear: both;
-  display: table;
-}
-
-
-@media screen and (max-width: 300px) {
-  .cancelbtn, .signupbtn {
-     width: 100%;
-  }
-}
-</style>
-<body>
-
-<form action="log.html" style="border:1px solid #ccc">
-  <div class="container">
-   
-    <p>Please fill in this form to create account with us</p>
-    <hr>
-
-    <label for="email"><b>Email</b></label>
-    <input type="text" placeholder="Enter Email" name="email" required>
-    <label for="email"><b>username</b></label>
-    <input type="text" placeholder="Enter your name" name="usernames" required>
-    <label for="psw"><b>Password</b></label>
-    <input type="password" placeholder="Enter Password" name="psw" required>
-
-    <label for="psw-repeat"><b>Repeat Password</b></label>
-    <input type="password" placeholder="Repeat Password" name="psw-repeat" required>
-    
-    <button type="submit" class="signupbtn">Sign Up</button>
-    <button type="submit"><a href="ui.html">cancel</a></button>
-    
-  
-  </div>
-</form>
-
-</body>
-</html>
