--- UI/html/ui.html
@@ -1,114 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="X-UA-Compatible" content="ie=edge">
-<link rel="stylesheet" href="../css/style.css">
-<title>Document</title>
-</head>
-<body>
-
-<header class="header">
-
-<nav>
-<ul class="nav-menu">
-<img class="logo"src="../assets/logo.png">
-
-<li> <a href="../html/about.html" > About us</a> </li>
-<li> <a href="../html/log.html" > login</a> </li>
-<li> <a href="../html/signup.html" > signup</a> </li>
-</ul>
-</nav>
-</header> 
-<div class="content">
-    
-    <marquee direction="right"><p> WELCOME TO MY DIARY</p></marquee>
-    <div class="content-b">
-<p> this is where you can keep your story the way you want<br>
-    if you want to start your Dairy<br>
-     <a href="log.html">CLICK HERE</a>
-</p>
-</div>
-   </div>
-
-</div><br><br><br><br><br><br>
-<div class="main-content">
-     
-</div>
-<div class="articles">
-        <div class="article content1">
-        
-      
-    </div>
-
-       
-    
-<div class ="footer">
-    
-    
-    
-</div>
-
-</div>
-
-    
-    <div class="kiki" >
-        
-        <h3> Testimony</h3>
-        <div class="articles">
-            <div class="article content1">
-          <img src="../assets/1.jpg"  width="140px" height="200px">
-        </div>
-
-            <div class="article content2">
-                <img src="../assets/2.jpg"  width="140px" height="200px">
-        </div>
-        <div class="article content2">
-            <img src="../assets/3.jpg"  width="140px" height="200px">
-        </div><br><br>
-        <div class="kiki" >
-        
-            
-            <div class="articles">
-                <div class="article content1">
-                
-                <p> using diary is something important<br>
-                    in our life , it help us to keep memory<br>
-                    by my side I can say that its really important.
-                </p>
-            </div>
-    
-                <div class="article content2">
-                
-                <p> using my diary is very good because <br>
-                it helps you to remember  important things<br>
-                which happened into life any time.
-             </p>
-            </div>
-            <div class="article content2">
-                
-                <p> I always use Diary to keep my secret<br>
-                    though its something bad sometimes beause <br>
-                    you might keep it in a wrong place and people<br>
-                    steel it .
-                </p>
-            </div>
-                   
-               
-    </div>
-
-    
-    
-        
-        
-
-    </div>
-    <div class="footer2">
-        <h2> copyright </h2>
-        <p> Emmanuella ,2019</p>
-       </div>
-   
-
-</body>
-</html>
