+++ package.json
@@ -7,7 +7,8 @@
     "test": "mocha --require @babel/register ./api.js",
     "build": "babel server -d build",
     "start": "nodemon -r esm ./server/index.js",
-    "dev": "nodemon --exec babel-node ./server/index.js"
+    "dev": "nodemon --exec babel-node ./server/index.js",
+    "create": "node User-Interface/server/config/connect.js createTables"
   },
   "repository": {
     "type": "git",
@@ -23,7 +24,10 @@
     "body-parser": "^1.19.0",
     "dotenv": "^8.1.0",
     "esm": "^3.2.25",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "make-runnable": "^1.3.6",
+    "nodemon": "^1.19.3",
+    "pg": "^7.12.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.6.0",
