--- server/controllers/all.js
@@ -1,8 +1,0 @@
-import data from '../models/db';
-const addall = (req,res) => {
-    res.status(200).json({
-        status:200,
-        data: data,
-    });
-}
-export default addall;
