--- server/middlewares/validator.js
@@ -1,7 +1,0 @@
-const validateEntry = (req, res, next) => {
-    const {title} = req.body;
-    if (!title.trim()) {return res.send('error')}
-    next();
-}
-
-export {validateEntry};
