--- server/models/db.js
@@ -1,23 +1,0 @@
-const data = [
-    {id: 1, 
-      title: 'my success', 
-       completed: false,
-        createdOn: new Date()},
-    {id: 2, 
-      title: 'my journey ',
-        completed: false, 
-        createdOn: new Date()},
-    {id: 3,
-       title: 'worst mistake',
-        completed: true,
-         createdOn: new Date()},
-    {id: 4,
-       title: 'happy', 
-        completed: false,
-         createdOn: new Date()},
-    {id: 5, 
-      title: 'sad',
-       completed: true, 
-       createdOn: new Date()},
-];
-export default data;
